backend/requirements/base.txt
Django==5.0.*
djangorestframework==3.15.*
djangorestframework-simplejwt==5.3.*
django-cors-headers==4.3.*
django-storages[s3]==1.14.*
boto3==1.34.*
celery==5.3.*
redis==5.0.*
anthropic==0.26.*
openai==1.*
stripe==9.*
python-decouple==3.8
python-magic==0.4.*
pypdf==4.*
python-docx==1.*
psycopg2-binary==2.9.*
dj-database-url==2.*
openai==2.21.0 
